Refactor web progress app data modules
