+++ apps/web/src/appData/index.tsx
@@ -3,4 +3,4 @@ export {
   useAppData,
 } from "./context/provider";
 export type { AppDataContextValue } from "./context/types";
-export { useReviewProgressBadge } from "./progress/reviewProgressBadge";
+export { useReviewProgressBadge } from "./progress/badge/reviewProgressBadge";
