+++ apps/web/src/appData/progress/badge/reviewProgressBadge.ts
@@ -1,8 +1,8 @@
 import { useMemo } from "react";
-import type { ProgressSummarySnapshot, ProgressSummarySourceState, ReviewProgressBadgeState } from "../../types";
-import { useProgressInvalidationState } from "./progressInvalidation";
-import { useProgressSource } from "./progressSource";
-import { useAppData } from "../context/provider";
+import type { ProgressSummarySnapshot, ProgressSummarySourceState, ReviewProgressBadgeState } from "../../../types";
+import { useProgressInvalidationState } from "../invalidation/progressInvalidation";
+import { useProgressSource } from "../progressSource";
+import { useAppData } from "../../context/provider";
 
 const EMPTY_REVIEW_PROGRESS_BADGE_STATE: ReviewProgressBadgeState = {
   streakDays: 0,
